Circle Test Draft
Goal:
Find out if I can use an Object of Point as a parameter of Circle. It's possible that I instead need to make individual parameters for each Point's x and y values under the Circle class. Current ideal is I can compare a Point object with the Circle's Point property, but I'm starting to feel like that wouldn't work.
Further testing required.